Implement inverse lookup
This patch augments the filterByParent logic to introspect Mirage’s ORM
layer in search of the `inverse` option on a hasMany relationship.

This allows us to handle cases where the key for the belongs to side of
an association is something other than the singularized name of the type
it’s pointing to.

For example:

```graphql
type Customer {
  comments: CommentConnection!
}

type CommentConnnection {
  edges: [CommentEdge]
}

type CommentEdge {
  cursor: String!
  node: Comment
}

type Comment {
  author # typically this would be "customer" but in this case it‘s more specialised
}
```